updated date use in test, repackaged, refactored code, added renamed @Indexed to unique db field and DirtiesContext in tests
