+++ .gitignore
@@ -34,3 +34,7 @@ build/
 /mvnw
 /mvnw.cmd
 /.mvn/wrapper/maven-wrapper.properties
+/et --hard 17bf7d3
+/tewiik 139918615+stewiik@users.noreply.github.com
+/itory, spring data, unit tests implementations
+/tec19@gmail.com
