+++ src/integration/java/pl/lotto/BaseIntegrationTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
@@ -18,10 +19,11 @@ import pl.lotto.domain.AdjustableClock;
 
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
 
+@DirtiesContext
 @SpringBootTest(classes = {LottoSpringBootApplication.class, IntegrationConfig.class})
-@ActiveProfiles("integration") // spring uzyje profil o tej nazwie - application-integration
-@AutoConfigureMockMvc //pozwoli zrobic klienta ktory zadaje zapytania do bazy
-@Testcontainers //pozwoli odpali baze na potrzeby testow na dockerze
+@ActiveProfiles("integration") // spring use a profile with this name - application-integration
+@AutoConfigureMockMvc // allow to make a client that queries the db
+@Testcontainers // allow to start the db for testing on docker
 public class BaseIntegrationTest {
 
     public static final String WIRE_MOCK_HOST = "http://localhost";
@@ -38,7 +40,7 @@ public class BaseIntegrationTest {
     @Container
     public static final MongoDBContainer monoDBContainer = new MongoDBContainer(DockerImageName.parse("mongo:4.0.10"));
 
-    @RegisterExtension //mini serwer na potrzeby testow integracyjncyh
+    @RegisterExtension // mini server for integration tests
     public static WireMockExtension wireMockServer = WireMockExtension.newInstance()
             .options(wireMockConfig().dynamicPort())
             .build();
