+++ src/main/java/pl/lotto/domain/numbergenerator/RandomNumbersGenerable.java
@@ -1,6 +1,8 @@
 package pl.lotto.domain.numbergenerator;
 
 
+import pl.lotto.domain.numbergenerator.dto.SixRandomNumbersDto;
+
 public interface RandomNumbersGenerable {
 
     SixRandomNumbersDto generateSixRandomNumbers(int count, int lowerBand, int upperBand);
