+++ src/main/java/pl/lotto/domain/numbergenerator/WinningNumbers.java
@@ -2,6 +2,7 @@ package pl.lotto.domain.numbergenerator;
 
 import lombok.Builder;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.time.LocalDateTime;
@@ -9,5 +10,5 @@ import java.util.Set;
 
 @Builder
 @Document
-public record WinningNumbers(@Id String id, Set<Integer> winningNumbers, LocalDateTime date) {
+record WinningNumbers(@Id String id, Set<Integer> winningNumbers, @Indexed(unique = true)LocalDateTime date) {
 }
