+++ src/main/java/pl/lotto/infrastructure/resultannouncer/controller/ResultAnnouncerRestController.java
@@ -1,5 +1,9 @@
 package pl.lotto.infrastructure.resultannouncer.controller;
 
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -7,6 +11,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 import pl.lotto.domain.resultannouncer.ResultAnnouncerFacade;
 import pl.lotto.domain.resultannouncer.dto.ResultAnnouncerResponseDto;
+import pl.lotto.infrastructure.resultannouncer.controller.error.ResultAnnouncerErrorResponse;
 
 @RestController
 @AllArgsConstructor
