+++ src/test/java/pl/lotto/domain/numbergenerator/WinningNumbersGeneratorTestImpl.java
@@ -1,5 +1,7 @@
 package pl.lotto.domain.numbergenerator;
 
+import pl.lotto.domain.numbergenerator.dto.SixRandomNumbersDto;
+
 import java.util.Set;
 
 public class WinningNumbersGeneratorTestImpl implements RandomNumbersGenerable {
